+++ test/cookbooks/test/recipes/repository.rb
@@ -1,0 +1,1 @@
+rundeck_repository 'public'