+++ test/fixtures/cookbooks/create_chef_dir/metadata.rb
@@ -1,2 +1,0 @@
-name 'create_chef_dir'
-version '0.0.1'