+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+# Unreleased
+* Complete rewrite of cookbook using custom resources
+* Adds circleci testing
+
 ## 4.2.0
 * Add optional LDAP data bag support
 * Support Rundeck 2.9.x